Move `VolumeAwareHitSampleInfo` to own file
